Removed redundant RFQ interaction tests
